Refactor config_flow and restore support for yaml configuration
