+++ custom_components/tauron_amiplus/const.py
@@ -5,6 +5,7 @@ DOMAIN = "tauron_amiplus"
 DEFAULT_NAME = "Tauron AmiPlus"
 DATA_TAURON_CLIENT = "data_client"
 CONF_METER_ID = "energy_meter_id"
+CONF_GENERATION = "generation"
 CONF_SHOW_GENERATION = "show_generation_sensors"
 CONF_URL_SERVICE = "https://elicznik.tauron-dystrybucja.pl"
 CONF_URL_LOGIN = "https://logowanie.tauron-dystrybucja.pl/login"
